requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
uvicorn
psycopg2-binary
passlib[bcrypt]
pyjwt
pwdlib[argon2]
python-jose
python-multipart